Migrating Graph and dfs to typescript and typescript support to the build system.
