+++ lib/alg/preorder.js
@@ -1,4 +1,4 @@
-var dfs = require("./dfs");
+var dfs = require("../../compiled/lib/alg/dfs.js").dfs;
 
 module.exports = preorder;
