+++ lib/alg/prim.js
@@ -1,5 +1,5 @@
 var _ = require("../lodash");
-var Graph = require("../graph");
+var Graph = require("../../compiled/lib/graph.js").Graph;
 var PriorityQueue = require("../data/priority-queue");
 
 module.exports = prim;
