+++ lib/index.js
@@ -1,5 +1,5 @@
 // Includes only the "core" of graphlib
 module.exports = {
-  Graph: require("./graph"),
+  Graph: require("../compiled/lib/graph.js").Graph,
   version: require("./version")
 };
