+++ lib/json.js
@@ -1,5 +1,5 @@
 var _ = require("./lodash");
-var Graph = require("./graph");
+var Graph = require("../compiled/lib/graph.js").Graph;
 
 module.exports = {
   write: write,
