+++ package.json
@@ -18,6 +18,7 @@
     "lodash": "4.17.21"
   },
   "devDependencies": {
+    "@tsconfig/recommended": "^1.0.1",
     "benchmark": "2.1.4",
     "browserify": "16.5.1",
     "chai": "4.2.0",
@@ -37,6 +38,7 @@
     "seedrandom": "3.0.5",
     "semver": "7.3.2",
     "sprintf": "0.1.5",
+    "typescript": "^4.6.3",
     "uglify-js": "3.10.0"
   },
   "repository": {
