+++ tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@tsconfig/recommended/tsconfig.json",
+  "compilerOptions": {
+    "outDir": "compiled/lib",
+    "noImplicitAny": false
+  },
+  "include": ["lib/**/*"],
+  "exclude": ["node_modules", "test/**/*"]
+}
